Fix output rounding errors resulting in black images and segfaults/OpenCL errors at full-resolution

+++ src/develop/pixelpipe_hb.h
@@ -292,10 +292,10 @@ void dt_dev_pixelpipe_synch_top(dt_dev_pixelpipe_t *pipe, struct dt_develop_t *d
 
 // process region of interest of pixels. returns 1 if pipe was altered during processing.
 int dt_dev_pixelpipe_process(dt_dev_pixelpipe_t *pipe, struct dt_develop_t *dev, int x, int y, int width,
-                             int height, float scale);
+                             int height, double scale);
 // convenience method that does not gamma-compress the image.
 int dt_dev_pixelpipe_process_no_gamma(dt_dev_pixelpipe_t *pipe, struct dt_develop_t *dev, int x, int y,
-                                      int width, int height, float scale);
+                                      int width, int height, double scale);
 
 // disable given op and all that comes after it in the pipe:
 void dt_dev_pixelpipe_disable_after(dt_dev_pixelpipe_t *pipe, const char *op);
